+++ EnhanceQoL/NewSettingsTable.lua
@@ -1,24 +1,15 @@
 local addonName, addon = ...
 
 addon.variables.NewVersionTableEQOL = {
-	-- 11.3.0-beta3 Focus Marker Macro
-	groupToolsFocusMarkerPreserveExisting = true,
-	groupToolsFocusMarkerUseMouseover = true,
-	landingPageButtonAnchor = true,
-	landingPageButtonCustomPosition = true,
-	landingPageButtonOffsetX = true,
-	landingPageButtonOffsetY = true,
-	landingPageButtonScale = true,
-	durationText = true,
-	chatIMFontSize = true,
-	UFGlobalAuraIgnoreMatrix = true,
-
 	--@eqol-beta@
-	-- 11.2.0-beta1 Mythic+
+	-- 11.4.0-beta1
 	mythicPlusTimerEnabled = true,
-
-	-- 11.3.0-beta3 Default Aura Containers
 	skinnerDefaultBuffIconsEnabled = true,
 	skinnerDefaultDebuffIconsEnabled = true,
+
+	-- 11.4.0-beta2
+	hideMinimapButtonBinToggle = true,
+	mythicPlusBloodlustTrackerGlowOnActive = true,
+	mythicPlusBloodlustTrackerShowActiveDuration = true,
 	--@end-eqol-beta@
 }